backend/requirements.txt
supabase>=2.0.0
fastapi==0.110.1
uvicorn==0.25.0
pydantic==2.12.4
python-dotenv==1.2.1
bcrypt==4.1.3
email-validator==2.3.0
websockets>=13.0
